+++ R/retractionwatch.R
@@ -1,6 +1,6 @@
 #' RetractionWatch data
 #'
-#' DOIs and nature of statements from the RetractionWatch database (downloaded 2025-02-28)
+#' DOIs and nature of statements from the RetractionWatch database (downloaded 2025-05-21)
 #'
 #' @format A data frame with 44784 rows and 2 columns:
 #' \describe{