+++ inst/modules/retractionwatch.R
@@ -1,0 +1,52 @@
+#' RetractionWatch
+#'
+#' @description
+#' Flag any cited papers in the RetractionWatch database
+#'
+#' @author Lisa DeBruine
+#'
+#' @references
+#' The Retraction Watch Database [Internet].
+#' New York: The Center for Scientific Integrity. 2018.
+#' ISSN: 2692-4579. [Cited 2025-05-20].
+#' Available from: http://retractiondatabase.org/.
+#'
+#' @param paper a paper object or paperlist object
+#' @param ... further arguments (not used)
+#'
+#' @returns a list with table, traffic light, and report text
+#'
+#' @examples
+#' module_run(psychsci, "retractionwatch")
+retractionwatch <- function(paper) {
+  # detailed table of results ----
+  refs <- papercheck::concat_tables(paper, c('references'))
+  table <- dplyr::inner_join(refs, papercheck::retractionwatch, by = 'doi')
+  if (nrow(table) > 0) {
+    cites <- papercheck::concat_tables(paper, c('citations'))
+    table <- dplyr::left_join(table, cites, by = c('id', 'bib_id'))
+  }
+
+  # summary output for paperlists ----
+  summary_table <- dplyr::count(table, id, retractionwatch) |>
+    tidyr::pivot_wider(names_from = retractionwatch, names_prefix = "rw_",
+                       values_from = n)
+
+  # determine the traffic light ----
+  tl <- if (nrow(table)) "yellow" else "green"
+
+  # report text for each possible traffic light ----
+  report <- c(
+    yellow = "You cited some papers in the Retraction Watch database (as of 2025-05-20). These may be retracted, have corrections, or expressions of concern.",
+    green = "You cited no papers in the Retraction Watch database (as of 2025-05-20)"
+  )
+
+  # return a list ----
+  list(
+    table = table,
+    summary = summary_table,
+    na_replace = 0,
+    traffic_light = tl,
+    report = report[[tl]]
+  )
+}