+++ inst/modules/statcheck.R
@@ -1,14 +1,55 @@
-stat_table <- papercheck::stats(paper)
+#' StatCheck
+#'
+#' @description
+#' Check consistency of p-values and test statistics
+#'
+#' @references
+#' Nuijten M, Epskamp S (2024). _statcheck: Extract Statistics from Articles and
+#' Recompute P-Values_. R package version 1.5.0,
+#' <https://CRAN.R-project.org/package=statcheck>.
+#'
+#' @import dplyr
+#'
+#' @param paper a paper object or paperlist object
+#' @param ... further arguments (not used)
+#'
+#' @returns a list with table, traffic light, and report text
+#'
+#' @examples
+#' module_run(psychsci[[10]], "statcheck")
+statcheck <- function(paper) {
+  # detailed table of results ----
+  stat_table <- papercheck::stats(paper)
+  table <- stat_table[stat_table$error, ]
 
-if (nrow(stat_table) == 0) {
-  tl = "na"
-} else if (all(stat_table$error == "FALSE") ) {
-  tl = "green"
-} else {
-  tl = "red"
+  # summary output for paperlists ----
+  summary_table <- dplyr::summarise(stat_table,
+                                    stats_found = dplyr::n(),
+                                    stats_error = sum(error),
+                                    decision_error = sum(decision_error),
+                                    .by = id)
+
+  # determine the traffic light ----
+  tl <- dplyr::case_when(
+    nrow(stat_table) == 0 ~ "na",
+    all(stat_table$error == "FALSE") ~ "green",
+    .default = "red"
+  )
+
+  # report text for each possible traffic light ----
+  report <- c(
+    na = "No test statistics were detected",
+    red = "We detected possible errors in test statistics",
+    green = "We detected no errors in test statistics",
+    fail = "StatCheck failed"
+  )
+
+  # return a list ----
+  list(
+    table = table,
+    summary = summary_table,
+    na_replace = 0,
+    traffic_light = tl,
+    report = report[[tl]]
+  )
 }
-
-list(
-  table = stat_table[stat_table$error, ],
-  traffic_light = tl
-)